Make sure that u-boot will store bootcount in environment.
Using default u-boot configuration bootcount is stored in RTC block.
Having this in softreset save register prevents from saving environment
on reboots, but it is not possible to change this value from userspace.

Changing the place of storing bootcount causes that fw_setenv utility
can be used to change its value in similar way as other boot variables.