MEN-307: BBB: Don't fall back on internal flash if boot fails.
Instead reset and retry from the memory card. Otherwise we do not fall
back to the previous partition, which is both important for testing,
and it is what Mender is supposed to do.

This removes the safety net of falling back to the internal flash
altogether, but I think this is a separate concern that we can tackle
later, perhaps when we have a bit more clarity on what Mender should
do in such a situation.

Signed-off-by: Kristian Amlie <kristian.amlie@mender.io>