Add SDIMG_DATA_PART_SIZE_MB variable.
This requires changing how we construct the data partition, since the
--size parameter for the partition doesn't do what you'd expect:
Instead of using it as the size of the partition, it adds it to the
size of the partition. We are pretty used to constructing our own
file systems now, so do that instead.

Also change the names of temporary file system files to be more
consistent.

Signed-off-by: Kristian Amlie <kristian.amlie@mender.io>