mender: workaround go install
Workaround go install inconsistency when installing host or cross
compiled binaries. Runnin go install, the built packages end up only in
${GOPATH}/bin or ${GOPATH}/bin/${GOOS}_${GOARCH}, the latter case being
true only for cross-compiled (from go's perspective) builds.

Guessing if we're go is cross-compiling is too much overhead, assume
that we're always cross-compiling. This works for the general case, but
will break if target is the same arch as host.

Signed-off-by: Maciej Borzecki <maciej.borzecki@open-rnd.pl>