MEN-440: Make storage device and partition numbers configurable.
The base-files recipe in particular needs to be aware of whether we
are building a mender image or not, which is why the mender-image
OVERRIDE is introduced. If this is used in the configuration,
base-files will use our fstab, whereas if not, it will not do
anything.

We also insert the selected information into other locations
appropriate for the build.

Signed-off-by: Kristian Amlie <kristian.amlie@mender.io>