Make it possible to use 'file://' URIs even in externalsrc builds.
The problem is that externalsrc causes *all* source fetching to be
skipped, even files that are local. These files are not mirrored in
the mender repository, so it makes it quite awkward to work it. Adding
manual copy rules for those files is too error prone, since we're
likely to add more files later.

So instead, do custom fetching of all files in SRC_URI that
specifically start with 'file://'.