Make U-Boot mender_altbootcmd return.
Previously it would go straight to bootcmd, but do this in the default
altbootcmd instead, so that mender_altbootcmd returns. This allows the
user to do additional, or alternative, operations in the remainder of
altbootcmd if he chooses to customize it.

Since both mender_altbootcmd and bootcmd will now run mender_setup, it
may run twice in the case of altbootcmd, but this has no harmful
effects, and is a small price to pay for the flexibility it offers.

Signed-off-by: Kristian Amlie <kristian.amlie@mender.io>