Add "mender_try_to_recover" boot command.
The intention behind it is to call it after normal boot has failed,
but before alternative boot methods are attempted, such as network
boot. This allows Mender to try to roll back instead, if possible,
which it does by rebooting and following the normal bootlimit
procedure.

If there is no upgrade in progress, the call does nothing and U-Boot
will carry on to try alternative boot methods, if available.