Split generic U-Boot patch into payload and integration pieces.
The motivation behind this patch is to separate the pure Mender code,
which is a patch that will never conflict (it only contains new
files), from the integration of that code into U-Boot. Even though the
latter patch is minimal, it may conflict with some U-Boot versions,
since it has to touch some core areas.

If there ever is such a conflict during a user's integration, he may
have to produce his own patch to get around the problem, but that
means he would have to duplicate all the Mender code, which is
unnecessary and also means he won't receive updates to that code
unless he manually ports it.

With the two patches, the payload, which is the Mender code, is
separate, and thus will stay up to date even if the user has made his
own integration patch.

Signed-off-by: Kristian Amlie <kristian.amlie@mender.io>