Remove file based update test.
Exactly the same functionality is tested by the network based test,
except for the small detail of reading from a file. However, reading
from a file is tested elsewhere, such as the broken update test.

The motivation behind removing the test is that it takes a long time,
and is highly redundant for the above reasons.

Signed-off-by: Kristian Amlie <kristian.amlie@mender.io>