Fix build for U-Boot 2016.11 and later.
Based on following user feedback:

Tested meta-mender on U-boot 2016.11 and get following:
| Error: You must add new CONFIG options using Kconfig
| The following new ad-hoc CONFIG options were detected:
| CONFIG_MENDER_H
|
| Please add these via Kconfig instead. Find a suitable Kconfig
| file and add a 'config' or 'menuconfig' option.

This is because of [PATCH 1/2] Generic boot code for Mender.

It is probably not the best naming convention to begin a header file
name with CONFIG_ since this is obviously now interpreted as
configuration option and hence the error.

Signed-off-by: Kristian Amlie <kristian.amlie@mender.io>