Reduce default size of boot partition to 16MB.
The environment is not stored there anymore, so we should keep it
minimal size by default.

I wanted to reduce it to the alignment size, or 8MB, but this didn't
work for some reason, and created a boot partition that was slightly
bigger and not aligned, despite containing no files.

Signed-off-by: Kristian Amlie <kristian.amlie@mender.io>