Remove test which corrupts the oldest of the two redundant U-Boot envs.
This will never happen in practice, and the test breaks if you had a
preceding test which left a corrupted environment in the oldest of the
two (which the test itself would do, hence it will not pass twice in a
row). While that could be fixed, remove the test instead, since it's
not a relevant real life scenario.

Signed-off-by: Kristian Amlie <kristian.amlie@mender.io>