demo: netbase: add option to specify address of Mender server
This will add an option so that the user can specify the Mender server
IP address during build time.

e.g. in local.conf:

	MENDER_DEMO_SERVER_IP_ADDRESS = "192.168.0.100"

This will append /etc/hosts with an entry linking
MENDER_DEMO_SERVER_IP_ADDRESS to docker.mender.io and
s3.docker.mender.io.

This will fix following error:

	cannot validate certificate for 192.168.1.20 because it doesn't
	contain any IP SANs

Which is related to the fact the demo certificate is for
docker.mender.io domain only.

If not set, /etc/hosts is untouched.

Signed-off-by: Mirza Krak <mirza.krak@gmail.com>