change MENDER_BOOT_PART to optional
It is un-common for other storage medias other then eMMC/SD to have an
boot part. e.g RAW flash, there the bootloader is usually embedded at
the start of the flash and this part is only accessible trough MTD
devices which is not really in the scope of what MENDER_BOOT_PART
defines (it is not part of a partition/volume layout).

We remove the fstab entry of MENDER_BOOT_PART if it is not specified,
otherwise this could cause problems when the system boots.

We print a warning if MENDER_BOOT_PART is not set, just to bring it to
users attention and make sure that they are doing this intentionally.

Changelog: Title

Signed-off-by: Mirza Krak <mirza.krak@gmail.com>