systemd_%.bbappend: Using += to add features to PACKAGECONFIG doesn't work.
PACKAGECONFIG is set in the systemd recipe using ??=. The += in the
bbappend sets the PACKAGECONFIG to "resolved networkd". Since the variable
is now set the ??= isn't applied. Since the _append runs after the ??= is
processed, use this method to append to the PACKAGECONFIG variable.

Signed-off-by: Philip Balister <philip@balister.org>

Changelog: Fix meta-mender overriding PACKAGECONFIG instead of adding
to it.

Modified-by: Kristian Amlie <kristian.amlie@mender.io>
 * Added missing space in variable definition.

Signed-off-by: Kristian Amlie <kristian.amlie@mender.io>