MEN-1014: Add mender-state-scripts class and example.
The new mender-state-scripts class can be used in a recipe to turn
that recipe into a state script recipe. Any script or binary that the
recipe provides using the class will become part of the Mender update
as a state script.

There are several ways to include them:

* Put the script/binary in ${MENDER_STATE_SCRIPTS_DIR} during the
  do_compile stage.

* Put an archive in SRC_URI that extracts to
  ${PN}-${PV}/mender-state-scripts, where ${PN}-${PV} are the recipe
  name and version.

* Set the variable MENDER_STATE_SCRIPTS manually to a directory that
  contains the scripts.

Note that the three methods should not be mixed.

Changelog: Commit

Signed-off-by: Kristian Amlie <kristian.amlie@mender.io>
(cherry picked from commit ea1261106deaf325db6437c409abb738203385ed)