meta-mender-qemu/vexpress-qemu-flash: add vexpress-nor image
Add support for generating images suitable for use as NOR flash for vexpress-a9
machine. The output image is actually a tar of 2 separate images for NOR banks 0
and 1 (named nor0 and nor1 respectively). The output file is named
`${IMAGE_NAME}.vexpress-nor` and is created automatically when building for
vexpress-qemu-flash.

Changelog: title

Signed-off-by: Maciej Borzecki <maciej.borzecki@rndity.com>