move systemd bbappend to meta-mender-demo
Bringing in meta-mender-core should make minimal changes to features
that are considered distribution features. In this case we should not
try and enforce usage of networkd just because Mender is used, since it
does not really have a hard dependency on it and this is better suited
for the meta-mender-demo layer.

Changelog: None

Signed-off-by: Mirza Krak <mirza.krak@endian.se>
(cherry picked from commit 5e1e85d5f1f97878aba34934890565b6d199406a)