mender: Do not build rpi-sdimg FSTYPE for Raspberry Pi.
This no longer works with the Mender customizations so recommend to
the user that they disable it.

Also note that this requires the changes from:
     https://github.com/drewmoseley/meta-mender/commit/c41e5cd1af651efffda7c2840dda0bacd00e3f68
as the old style IMAGE_DEPENDS_* variable does not seem to work properly
and the SDIMG build fails missing the WIC tools without it.

Changelog: Title

Signed-off-by: Drew Moseley <drew.moseley@northern.tech>