tests/meta-mender-ci: add mender-image-test-host
Add an image for QA test host. Test host is a machine that will be running QA
tests in a CI pipeline. Test host image is a rootfs that the machine will boot
into by default. When executing on-board QA tests, the CI process will log onto
the test host, upload the image under test, deploy to local storage and reboot
the host to run the new image. Once the tests have completed, the board will
return back to booting into test host image.

The test host image needs to have all the tools needed for performing CI tests
related to uploading and deploying the image + other programs that are needed
for establishing network connectivity and so on.

Changelog: none
Signed-off-by: Maciej Borzecki <maciej.borzecki@rndity.com>