meta-mender-core/u-boot: use weak assignment for MENDER_UBOOT_ENV_STORAGE_DEVICE_OFFSET_{1,2}
There may be a need to override the offsets manually. For instance, when the
board is set up with CONFIG_ENV_RANGE, the redundant environment will not be at
a BOOTENV_SIZE offset from the primary one, but rather at CONFIG_ENV_RANGE
offset. Instead of guessing, allow the user to configure the offsets as needed.

Changelog: title

Signed-off-by: Maciej Borzecki <maciej.borzecki@rndity.com>