tests/test_ubimg: account for large rootfs size
UBIFS does not store empty blocks, so if the rootfs is only modestly filled with
data, and the storage size is large, it's likely that the actual image size will
be relatively small compared to all available storage space.

Use simple heuristic, if rootfs size is < 100MB, expect the ubimg file size to
be > ~30% of total storage, otherwise ubimg shall be more than 10% of total
storage size.

Changelog: none

Signed-off-by: Maciej Borzecki <maciej.borzecki@rndity.com>