MEN-1293: Fix duplicate entries appearing in fstab.
wic does its own population of fstab in the generated sdimg if a path
is specified. It's tempting to utilize this and avoid populating it
ourselves, but we can't do that because it then the sdimg and the
.mender update containing the base ext4 would be different, since the
latter does not get this treatment. So instead remove the path from
the wic spec file to avoid the extra entries.

Changelog: Title

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>