meta-mender-qemu/mender-qemu: extract NOR flash data files to a temporary directory
mender-qemu will now create a temporary directory /tmp/mender-qemu.XXXXXXX to
store any files that need to be created during a particular run. In case of
vexpress-nor images, the nor* files will be extracted to the temp directory.

Set TMPDIR environment variable to override the path.

Since we need to do cleanup at the end (or when SIGTERM is received), the script
no longer execs qemu-system-arm.

Changelog: title
Signed-off-by: Maciej Borzecki <maciej.borzecki@rndity.com>