tests/common: when running mender-qemu put it in a separate process group
mender-qemu no longer execs qemu-system-arm, hence killing the script would
leave a running qemu instance behind. Make sure that mender-qemu is run in a
separate process group and that termination signals are sent to the whole group.

Make sure that the qemu cleanup is using terminate() to give mender-qemu a
chance to do the cleanup.

Changelog: none

Signed-off-by: Maciej Borzecki <maciej.borzecki@rndity.com>