Fix incorrect precedence in default assignments.
It turns out that assignments with overrides (like "_mender-ubi")
takes precedence over all other assignments, regardless of how weak
the assignment is. Since the intention is to provide a truly weak
assignment, we need to do this with a snippet instead, which is
protected by a feature guard.

Changelog: None

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>