MEN-1431: Make UEFI image generator: uefiimg.
This is the first step in booting via UEFI. This just creates an image
using the GPT partition table format, which UEFI requires. There are
some shortcomings in this implementation, which will be resolved in
upcoming tasks:

* The boot partition source in the wks file should be changed to
  'bootimg-efi'. However, this can't be done until we have a valid
  bootloader available like grub. Hence the EFI partition is not valid
  at the moment.

* Right now you have to set IMAGE_BOOT_FILES to a dummy file in order
  for wic not to complain. This will also go away once we get a real
  EFI bootloader in place.

We also tweak some variable assignments after it was revealed that
removing mender-uboot from DISTRO_FEATURES caused some problems.

Changelog: title

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>