Fix broken evaluation order when calculating variables.
It turns out that using getVarFlag() and "_defaultval" is not a good
method because bitbake is not able to calculate variable evaluation
order when using them. This leads to some variables being calculated
before the variables they depend on have received their final value,
leading to incorrect variable content.

Since we cannot assign default values with overrides (see b8b6e4640f),
we instead create a new set of variables that hold a default value,
and then set the final variables from these. This means that the
variables can still be changed successfully by users, since all the
overrides we use only apply to the variables that have a "_DEFAULT"
suffix, and hence the base variables do not use any overrides that
would take precedence over the users' assignments.

Changelog: Title

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>