u-boot: update mender_boot_part_name when mender_boot_part changes
mender_boot_part_name was never updated once it had been written from
the default environment.

Which led to that we never updated the Linux kernel and dtb on
a NAND/UBI setup, becuase we would always load the Linux kernel from
ROOTFS_A regardless of current active part.

While at it also removed the "persistant" storage of mender_boot_part_name
and the reason for this is that the persistant value would not match the
actual value used to boot the device since we alter it in "mender_setup"
which does not re-save the environment.

Changelog: Title

Reported-by: Dan Walkes <danwalkes@trellis-logic.com>
Signed-off-by: Mirza Krak <mirza.krak@endian.se>