Work around bug in libpseudo regarding file owners on data partition.
For some reason if we let rsync set the owner and group information
first, it becomes impossible to set them to something else later on.
The exact cause is unknown, but it may be that rsync uses fchown(),
while we use chown(), and that this triggers the bug. In any case,
avoiding setting the owner and group with rsync seems to fix the
problem.

Changelog: Title

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>
(cherry picked from commit b878227ead8c9b581448d1c1911fca9423b9828f)