Improve resiliency of RPi3 image flashing by booting Raspbian kernel.
Previously we were booting the Mender kernel, even when we were
heading for the Raspbian OS to do a Mender sdimg flash. With this
commit we create a separate boot.scr file that can be used to boot the
kernel directly from Raspbian, avoiding any problems that might occur
because some Mender test has corrupted the Mender kernel.

Changelog: None

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>