MEN-1387: Implement heuristic automatic patching of U-Boot.
The main tools in the process are uboot_auto_configure.sh and
uboot_auto_patch.sh. The former sets up the necessary prerequisites,
and the latter does the actual patching.

The prerequisites include making a copy of the source and then
compiling a host version of the fw tools. This gives access to a
direct dump of the default environment, as well as build information
that helps with locating files that need to be patched.

The last step in uboot_auto_configure.sh is calling
uboot_auto_patch.sh, which uses the information from the prerequisite
steps to go through all the required patch steps. This is basically
everything listed at [1].

As part of this, we delete the patch files for QEMU and Beaglebone,
since they are not needed anymore. QEMU needs a tiny new patch, which
is necessary only because we ourselves have created the vexpress-qemu
machine type, and we need to define load addresses for the kernel and
dtb file. Raspberry Pi unfortunately has to keep its manual patching
because of its use of the boot.scr and cmdline.txt features, which we
currently don't patch correctly. All the Flash variants are also left
alone for now.

Beaglebone gets a new CI specific patch, however, since we need to
reenable the uenvcmd feature that we use to switch booting from the
memory card to the internal storage. This addon functionality is not
provided by the auto-patching mechanism.

Since the u-boot-fw-utils-mender-auto-provided will now be used by
default, and it grabs its sources from whatever u-boot built, then we
also remove all fw-utils specific patches. Patches for the two
packages should always be identical anyway, so this is just following
good practice.

Testing this feature is tricky and time consuming. A couple of tests
have been added to test that the MENDER_UBOOT_AUTO_CONFIGURE variable
and other u-boot recipe related steps are working correctly. In
addition there is also a test which compile tests all U-Boot boards,
and this test is monstruous, taking more than an hour even on a 20
core machine. To deal with this, the test has revisions listed inside
it that can be used to skip the test for revisions that have already
been tested. Since the test only tests U-Boot source code and our
patching tools, it will only ever change status if either the U-Boot
version has changed, or one of our patching tools for U-Boot. Mender
source code changes have no effect on it.

[1] https://docs.mender.io/1.2/devices/integrating-with-u-boot

Changelog: Implement heuristic automatic patching of U-Boot.
It can be turned on and off by setting `MENDER_UBOOT_AUTO_CONFIGURE`
to `1` and `0`, respectively, in a `u-boot.bbappend` file. It is on by
default. If the automatic patching is unsuccessful, there is a special
bitbake target that can be used to extract the generated patch and use
it as a basis for a manual patch. It can be invoked with `bitbake -c
save_mender_auto_configured_patch <u-boot-recipe>`, where
`<u-boot-recipe>` is either `u-boot` or the fork of U-Boot that your
board uses.

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>