MEN-1387: Remove now unneeded meta-mender-beaglebone layer.
Originally the fix was to remove the custom BOOTENV_SIZE_beaglebone,
which already defaults to 0x20000, so the definition is redundant.
This causes the layer to become empty however, so the avalanche effect
is that we remove the whole layer.

Changelog: Remove meta-mender-beaglebone layer. This layer is not
needed anymore for compiling for Beaglebone and should be removed from
all build configurations.

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>