tests: Prevent printing of entire 8MB update.
After 6a420b51ef80ef, the update became much bigger, so the flipped
byte will always be outside the range of the content string we test
for, since it is in the middle of the file. So we no longer need to
update the content, and in fact doing so would print the entire thing
in the log.

Changelog: None

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>
(cherry picked from commit 42e072774d18788ca930e28631f49e0e33748b92)