meta-mender-core: Allow dtbos in KERNEL_DEVICETREE
Raspberry Pi kernel recipes use the KERNEL_DEVICETREE
variable to specify the dtbo overlays to build, not just
the board dtbs.

Ignore these dtbo overlays when assigning MENDER_DTB_NAME.

The existing behavior of choosing the last dtb in the list
if more then one is found is maintained.

Not sure this is a particularly good idea. It might be better
to abort if multiple dtbs are found since using the wrong
dtb for uboot will probably not work.

Signed-off-by: Scott Ellis <scott@jumpnowtek.com>

Modified-by: Kristian Amlie <kristian.amlie@northern.tech>
* Add quotes around variable assignment
* Switch to sed extended regexes and '\s+'
* Remove now unneeded "grep -o", and replace with whitespace stripping
* Remove redundant 'sed' expressions
* Add changelog

Changelog: Title

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>