tests: Move fixtures into own file to avoid multiple session fixtures.
If common.py is repeatedly imported from several tests, this causes
pytest to create a separate session fixture each time it is imported,
even though there should be only one. This reduces performance
significantly because for instance the test build directory needs to
be reinitialized several times. It can also be dangerous because it is
not expected that session fixtures should have more than one instance.

Fix it by moving the fixtures into their own file and importing it
only from conftest.py. The remaining functions are left in common.py,
since many of them need to be imported to be used as function calls
from the various tests.

Changelog: None

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>