Fix 'depends upon non-existent task' error in U-Boot recipes without auto patching.
The problem is this: When using the '[depends]' mechanism, the
mentioned ${PN} will be translated into the current provider of that
name. For example, if the recipe is u-boot-fw-utils, but the
'PREFERRED_PROVIDER_u-boot-fw-utils' is set to 'u-boot-my-fw-utils',
then bitbake will look for the task in the latter recipe. If this
recipe has set MENDER_UBOOT_AUTO_CONFIGURE to 0, then the task does
not exist, and we get the error from the title.

The solution is to move the dependency into the addtask() call
instead, since this correctly refers the dependency to the task in the
same recipe.

Changelog: Title

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>