Fix "No rule to make target 'envtools'" error in some U-Boot builds.
In u-boot-fw-utils-mender-auto-provided, we don't know what sources we
will be served. They can be both new and old. So we need to apply the
same fix that already exists in the uboot_auto_configure.sh script:
Detect which of "env" and "envtools" we need to use to build the host
tools.

Changelog: Title

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>