Added new state script to wait for time sync to complete.
This heuristically just checks the current year and if it is older than 2017
we can assume that it is not synced yet.  We will wait up to 1 minute, checking
every 5 seconds to let it settle before continuing to try and connect to
the server.

Changelog: Title
Signed-off-by: Drew Moseley <drew.moseley@northern.tech>
(cherry picked from commit e85b97fff3c08fc3579574c0ad2a95ce35020243)