Make sure that older "beaglebone" devices can upgrade to "beaglebone-yocto".
Maybe this should be in a BSP layer, but we will not create a whole
layer just for that. It is only a default value anyway.

Changelog: Since the "beaglebone" machine name has changed in upstream
to "beaglebone-yocto", add "beaglebone" to the
`MENDER_DEVICE_TYPES_COMPATIBLE` default, so that older devices can
upgrade even if they have the old device type.

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>