mender-uboot: Backport CONFIG_MENDER_BOOTCOMMAND.
This is part of the U-Boot autopatching code in rocko and newer.

For some platforms, We can generate the patch using the newer branch
and then manually apply it here in this branch.  That requires this
change to be effective.

Note that unlike in rocko and newer, where the application of:
  0004-Disable-CONFIG_BOOTCOMMAND-and-enable-CONFIG_MENDER_.patch
is conditional on the MENDER_UBOOT_AUTO_CONFIGURE setting, we will
only add that patch specifically to boards that will have their
U-Boot patch generated in this manner.

Changelog: Title
Signed-off-by: Drew Moseley <drew.moseley@northern.tech>