fix support of CONVERSIONTYPES for sdimg
The image_types.bbclass defines a set of conversion types that can be
applied to images (gz, xz, bz2 and more).

With the current state it is not possible to apply an conversion type
on the sdimg using:

	IMAGE_FSTYPES += "sdimg.gz"

Because the conversion types assume that that the image is named in the
following format:

	${IMAGE_NAME}${IMAGE_NAME_SUFFIX}.${type}

IMAGE_NAME_SUFFIX defaults to ".rootfs" which breaks things for us. By
setting IMAGE_NAME_SUFFIX to an empty string the expected image name
format is correct and the sdimg type can support all the conversiontypes
that are present in image_types.bbclass.

Above is really helpfull to compress the "sdimg" file as it mostly consist
of zero's and by compressing it is much easier to handle, transfer between
devices or between servers.

Note that this change does not generate a compressed sdimg by default, but
only adds support for it.

Changelog: None

Signed-off-by: Mirza Krak <mirza.krak@endian.se>
(cherry picked from commit d5cd89d708adec703907dce8616c5c735d0e5aba)