MEN-1597: Promote mtdids and mtdparts information to Bitbake variables
The motivation behind the change is allowing different parts of the
build process to be influenced by their contents. This was not
possible before because all the information resided in U-Boot source
code, which is not accessible outside of that recipe.

It's not completely automatic because one has to define mtdid of the
storage one wants to use. Maybe this can be automated further by
improving the auto patching from the sdimg, but even this might be
tricky, so we might settle for simply making it easy to extract the
information manually, and then use that. For now MENDER_MTDIDS has to
be set to exactly one mtdid assignment, and the rest is automatic, or
one can set it to multiple mtdids, but then MENDER_IS_ON_MTDID and
MENDER_MTDPARTS have to be defined manually.

Changelog: The handling of mtdids and mtdparts (`MTDIDS_DEFAULT` and
`MTDPARTS_DEFAULT` defines in U-Boot source code) has changed. The
defines now need to be patched out of U-Boot source code if they are
present, and added to `MENDER_MTDIDS` and `MENDER_MTDPARTS` bitbake
variables, respectively. The benefit is that build procedures that are
influenced by their values no longer need to be hardcoded, such as the
offset for the U-Boot environment, whose value will be automatically
determined. It will also be used in the upcoming mtdimg type.

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>