MEN-1597: Add mtdimg image type.
The mtdimg type is an image type meant to be flashed to the entire
Flash device, unlike the ubimg, which should only be flashed to the
ubi area of the mtd device. Either one can be used depending on need.

vexpress-nor image, used in QEMU, was also changed to take advantage
of the new mtdimg image.

Changelog: Commit

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>