mender-setup: add possibility to override Mender fstab entries
Currently one can only override the FSTYPE, but in some cases it
would be interesting to setup some application specific options as
well.

One example is the /uboot entry which mounts a vfat partition. vfat
does not suppoert file-permissions and by default it is mounted as
such:

    -rwxr-xr-x  1 root root  423 Feb 26  2018 random-file.txt

In a multi-user environment it is not possible write to this other
then with root permissions. vfat partitions are normally mounted
with "disk:disk" and rw permissions for that group. I can not
accomplish this in the current implementation

Changelog: replaced MENDER_DATA/BOOT_PART_FSTYPE with MENDER_DATA/BOOT_PART_FSTAB_ENTRY

Signed-off-by: Mirza Krak <mirza.krak@endian.se>