Replace/rename ALIGNMENT and RESERVED_RAW_SPACE variables.
Changelog: Replace `MENDER_PARTITION_ALIGNMENT_KB` with
`MENDER_PARTITION_ALIGNMENT`, which is in bytes.

Changelog: Rename `MENDER_STORAGE_RESERVED_RAW_SPACE` to
`MENDER_RESERVED_SPACE_BOOTLOADER_DATA`, to better reflect what it is
used for.

MENDER_PARTITION_ALIGNMENT is in preparation for UBI, where block
sizes are not generally a multiple of KiB.

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>