Rewrite partition logic and add Flash/UBI auto patching support for U-Boot
Changelog: Add U-Boot auto patching support for Flash based devices.
The autopatcher will use UBI as storage medium for both the
filesystems and the U-Boot environment. Mender requires a minimum of
configuration: the `MENDER_MTDIDS` needs to be set for the board, and
will normally go in the `conf/machine/<MACHINE>.conf` file for the
board in question. See documentation for variables `MENDER_MTDIDS`,
`MENDER_IS_IN_MTDID` and `MENDER_MTDPARTS` for more information.

Changelog: Partition alignment (`MENDER_PARTITION_ALIGNMENT`) on Flash
devices using UBI is now aligned to the UBI LEB size, which in general
is not a multiple of KiB.

Fixes MEN-1597.

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>