mender-qemu: More robust detection of MACHINE setting.
Determine the MACHINE setting from the bitbake conf files more closely aligned to
how bitbake does it.  There are multiple potential conf files as well as assignments
using '=' '?= and '??='.  The current implementation may miss some more esoteric
assignments but the only truly reliable way to figure it out is to run bitbake
and let it parse the metadata and I don't think we want to do that on every
invocation of this script.

Changelog: Title
Signed-off-by: Drew Moseley <drew.moseley@northern.tech>