mender-part-images: add ability to override /data partition's FS type
Add new ARTIFACTIMG_DATA_FSTYPE that can be used to specify FS different from the one
used for the rootfs. It defaults to the rootfs FSTYPE but can be overriden
if need be.

Changelog: add ability to override /data partition's filesystem to be
different from the one used for rootfs.

Signed-off-by: Georgii Staroselskii <georgii.staroselskii@emlid.com>