Implement IMAGE_ROOTFS_EXCLUDE_PATH support.
Changelog: Implement `IMAGE_ROOTFS_EXCLUDE_PATH` support. It works the
following way: It contains a space separated list of directories,
relative to the rootfs root (no beginning slash), and any directory
specified will be omitted from the rootfs. If the directory ends in a
slash, only the contents will be omitted, not the directory itself
(useful for mount points). One can then set
`do_image_<imagetype>[respect_exclude_path] = "1"` for certain image
builders to prevent the exclusion and then add them back as separate
partitions there.

Changelog: Support for `MENDER_DATA_PART_DIR` has been removed. Use
recipe files to add files directly to the `/data` directory instead.

Note that this variable does not begin with `MENDER_` and this is
intentional. Originally this was developed as an upstream feature in
OpenEmbedded, however due to their extremely limited capacity at the
time of writing, we have postponed upstreaming and added it in our
Mender layer instead. However, we still want to upstream it, and it
would be great if it continued using the same name in this case. The
feature has already been reviewed on OE's mailing list, so they should
be ok with both naming and behavior. TBD.

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>